requirements.txt
opencv-python
mediapipe
torch
torchvision
numpy
pygame
pillow